fix: ProductController 의 누락된 사용자 확인 일치 로직 추가
삭제자의 신원이 Product의 소유자 혹은 관리자 직책인지 검증

+++ com.3NT.deliveryShop/src/main/java/com/_NT/deliveryShop/controller/ProductController.java
@@ -125,6 +125,7 @@ public class ProductController {
     public Result.Deleted softDeleteProduct(@PathVariable UUID productId,
         Authentication authentication) {
 
+        productAuthorizer.requireProductOwner(authentication, productId);
         Result.Deleted productDeleteResult = service.softDeleteProduct(productId, authentication);
         productImgService.deleteProductImg(productId, authentication);
         return productDeleteResult;
